Import and Export work now, but I am chagning the output format a little and I will be changing the import code to accomidate it.

+++ Database/Query/queryFunctions.php
@@ -15,7 +15,8 @@ function generateUpdateQuery($tableName, $columns, $rowData)
     {
 	//Check if the data value has been put inside single-quotes and add
 	//them if needed
-	if($rowData[$i][0] != "'")
+	$len = strlen($rowData[$i]);
+	if( $len <= 1 || $rowData[$i][0] != "'" || $rowData[$i][$len-1] != "'" )
 	    $set = $set . $columns[$i] . "='" . $rowData[$i] . "'";
 	else
 	    $set = $set . $columns[$i] . "=" . $rowData[$i];
@@ -24,7 +25,8 @@ function generateUpdateQuery($tableName, $columns, $rowData)
 	    $set = $set . ",";
     }
 
-    if($rowData[0][0] != "'")
+    $len = strlen($rowData[0]);
+    if( $len <= 1 || $rowData[0][0] != "'" || $rowData[0][$len-1] != "'" )
 	return "UPDATE " . $tableName . " SET " . $set . " WHERE " . $columns[0] . "='" . $rowData[0] . "'";
     else
 	return "UPDATE " . $tableName . " SET " . $set . " WHERE " . $columns[0] . "=" . $rowData[0];
@@ -48,12 +50,22 @@ function generateInsertQuery($tableName,$columns,$values)
 	if($i > 0)
 	{
 	    $columnString = $columnString . "," . $columns[$i];
-	    $valueString  = $valueString . ",'" . $values[$i] . "'";
+	    
+	    $len = strlen($values[$i]);
+	    if( $len <= 1 || $values[$i][0] != "'" || $values[$i][$len-1] != "'" )
+		$valueString  = $valueString . ",'" . $values[$i] . "'";
+	    else
+		$valueString  = $valueString . "," . $values[$i];
 	}
 	else
 	{
 	    $columnString = $columns[0];
-	    $valueString  = "'" . $values[0] . "'";
+
+	    $len = strlen($values[0]);
+	    if( $len <= 1 || $values[0][0] != "'" || $values[0][$len-1] != "'" )
+		$valueString  = "'" . $values[0] . "'";
+	    else
+		$valueString  = $values[0];
 	}
     }
     
@@ -72,6 +84,10 @@ function generateInsertQuery($tableName,$columns,$values)
  */
 function generateDeleteQuery($tableName,$column,$value)
 {
-    return "DELETE FROM $tableName WHERE $column='$value'";
+    $len = strlen($value);
+    if( $len <= 1 || $value[0] != "'" || $value[$len-1] != "'" )
+	return "DELETE FROM $tableName WHERE $column='$value'";
+    else
+	return "DELETE FROM $tableName WHERE $column=$value";
 }
 ?>
